Extract pipeline artifact persistence from inference
