+++ tests/test_architecture_runtime_split.py
@@ -9,6 +9,7 @@ def test_internal_runtime_modules_stay_small():
         package_root / "batching.py",
         package_root / "distributed.py",
         package_root / "hierarchical.py",
+        package_root / "persistence.py",
         package_root / "progress_bridge.py",
         package_root / "serialization.py",
         package_root / "types.py",
